[lsp] Request Coq interrupt when we receive a new check request.
More testing for this is needed, but that's a nice first step, as
actually setting the flag will eventually raise an async exception
which is what we want in the general case.

Note however that we must indeed handle the async interruption in our
code properly, in particular we MUST NOT cache the user-interrupt,
that is to say, we need to make the memo cache clean up properly on
the async interrupt.