[serapi] Bump to version that implements hash / compare for tactics.
This makes the cache much more efficient, with now almost perfect hit
rate.

However, this also makes obvious the problem we have with our caching
system as indeed located messages (and diagnostics) are unsound
w.r.t. to location-independent caching.

We must implement a proper fix to this strategy before merging this; I
think we can indeed make the cache aware of the location a command was
cached, and thus include the offset when we get a hit, and adjust
the cached messages accordingly. Exciting!