[interp] Cache goals and feedbacks in the interpretation cache.
Goal printing is dependent on the state, thus it makes sense to cache
it, as otherwise we'd have to restore the state to print them, which
would kind of defeat the purpose of caching.

This fixes a bug when we'd print goals in the wrong state as `Memo`
wouldn't set the state correctly.

Feedbacks are not, per-se, dependent on the state, but we cache them
too as it makes sense to modularize the code that way.