[fleche] [controller] First steps towards resumption
As of today, checking a document can only be preempted by another
`didChanges` request. This is inconvenient when working in large
documents, as we would like goals or hover to be able to interrupt it.

We add support in the controller for preempting checking by
requests. The logic is that we keep a pointer for checking under
progress, which is only set to `None` when the fleche notifies the
checking has been completed. The queue manager will preempt checking
for all requests, and the queue is empty, resume checking if pending.

As of this PR, resuming checking will re-check the document in
full (with memoization of course); this is to be improved likely in a
different PR, maybe after parsing memo?

Note that this interacts with other desired features such as lazy /
viewrange checking, parsing memoization, and other features in
non-trivial ways.

Fixes #70