[fleche] Resumption of document checking
We now resume document checking from the point it was left at. That
required some fixing upstream, see https://github.com/coq/coq/pull/16978 .

It seems to work well and be simple enough.

This does stress `Coq.State.compare`, so I did implement a much weaker
version; unfortunately we don't have a test-suite for checking that
memo is working properly, so we need to keep an eye on it and do some
manual testing.