[fleche] Store diagnostics per node, create nodes for unparseable spans
We rework the document definition so now every span of the document
that we process will have a node, even if an Ast can't be produced.

This makes more sense and allows us to store the diagnostics per-node;
moreover, unparseable bits have now an associated node, which is
important for completion and other use cases such as common prefix
handling.