[fleche] Implement fileProgress notification for document progress
This is modelled after Lean's `$/lean/fileProgress` notification: a
yellow bar in the right side panel indicates what parts of the
document are still not checked.

Fixes #54

Note that we send on fileProgress per sentence, that's likely too
much; for now, we throttle the decoration update. Could be made into a
user-config.