[fleche] Remove hack for error recovery state
Up to now, we stored error recovery states in a global variable. This
doesn't play well with more advanced resumption setups.

We use a more robust approach (tho still far from final), and scan the
document back for a recovery candidate. This is much better, and
although it has a linear cost until we optimize it, should be much
more robust, and allow the implementation of #90 in a very easy way.