[infoview] Show Notice feedbacks on panel instead of as diagnostics
This seems more sensible as Coq (legacy) will produce large amounts of
these kind of Coq messages when for example `Search` or `About` is
used.

Thus, users can click on the concrete command and obtain the messages
information in the side panel; moreover, this allows for better
rendering opportunities.

We also implement a detailed display for errors, as Coq errors can be
large and don't fit properly in the diagnostics list.

A configuration option allows to restore old behavior.

Fixes #125