[logging] Implement `$/logTrace` according to spec.
It seems that #130 did a strange interpretation of the spec, the
`verbose` field is supposed to contain extra information, not a
`message`/`verbose` string.

Moreover, the logic in `logTrace` is strange, as the set of verbose
traces didn't contain the set of message ones.

We provide semantics more according to the protocol for now.