[controller] Postpone position requests until doc checking catches up.
This provides a much nicer typing and navigation experience, basically
if some data about the document is not ready, the request is postponed
and will be served when such data becames available.

We also switch the default for the client as to show goals on typing /
movement of the cursor. With the new delay capability, this gives the
user usually what they want in terms of the goal panel being up-to-date.

Some notes:

- For now we only postpone goal requests, we should implement a proper
  queue (in next PRs). One pending goal request works pretty well for
  interactive use.

- `Doc.Target.t` can be more refined, as to support `Info.find` `Prev`
  behavior (but this is a bit complex due to the way both searches interact)