[build] [code] [ci] Check TS compilation on CI.
Recent changes to our setup meant that compile is not type checking
anymore, thus CI didn't catch a type error that broke the extension.

We now test the release build in CI then, and maybe we could even
compare the output of `ls` to a set of files, and add some more
checks.

Thus, `make extension` is now really a "release" target.

(I don't want to make CI use `make`, but the default build driver for
 the VSCode marketplace)