[infoview] Port to React
This is prerequisite to implement / fix effectively #143 and #216 for
example; the current Info View rendering using strings doesn't scale
anymore to the kind of DOM manipulation we need.

Why React? That was a careful choice for many reasons, in particular
seems like the best fit for our use case (so far); moreover, 3rd party
package support is the best.