[doc] [completion] Improve UTF-8 and position validation
We are now more strict, and require the [Utf8] clients to take an
action when the conversion fails.

Usually clients doing the conversion for the range will want to select
the total length (last character) whereas the clients trying to grab a
character will want to abort whatever they are doing.

I have reviewed all such uses, and found a couple of errors.

This also fixes the bug with the `fileProgress` range due to the
above.

Re Fixes #264