[goals] Show better information about background goals.
This is an adaptation to React of the jsCoq goal printing code, but
our current model is a bit different so actually the code is not the
same.

In particular, we display bullets goals, and a bunch of extra
information, including goal preview.

Fixes #288

This is inspired by code originally written by Shachar Itzhaky.