[doc manager] Move postponement logic to document manager.
Document manager is the current owner of the build state, so it makes
sense that it is the component to decide whether a (data) request
should be postponed.

This fixes two code todos (`AbortRequest` and duplication of checks)
and should open the way for a deeper refactoring of `Doc_manager`
towards a true build system for a set of documents.