[compiler] Flèche-based command line compiler.
We expose Flèche compilation capabilities to the command line,
avoiding the need to use LSP for batch compiling.

The `fcc` Flèche Coq Compiler shares many of the features of the
Flèche-based LSP server, in particular, advanced error recovery,
markdown support, _CoqProject detection, caching (which is not yet
very exploited), and the possibility to access all metadata Flèche
knows about.

This PR is for now just a base to experiment, however it should work
fine for most files and use cases.