[vscode] Limit the number of messages displayed in the goal window
We limit the number of messages displayed in the goal window to 101,
to workaround slow render of Coq's pretty printing format.

This is an issue for example in Search where we can get thousand of
results.

We also speed up the rendering a bit by not hashing twice,
and fix a parameter-passing bug.

Thanks to Anton Podkopaev for reporting.