[lsp] Fix bug when retrieving URIs from document params
That made the `coq/saveVo` fail with an invalid parsing error.

The issue was that when we handle generically the requests that need
to wait for the document to be completed, we were retrieving the `uri`
parameter using the de-serializer for the `TextDocumentIdentifier`
data-type, however both `coq/getDocument` and `coq/saveVo` use
`VersionedTextDocumentIdentifier`.

As the JSON parser finds this extra param, it fails. We solve this
making the parser in this case no strict, but further hardening is
needed at this level.

Thanks to Théo Zimmerman for the report and help with debugging this
issue.