[fleche] Support for structure in documents
Nodes can now have children; this gives the consumers (error recovery,
document outline, analysis tools, etc...) a much more convenient
representation of the document.

We postpone adapting the document outline code to a further PR, so we
have tested this with the `coq/getDocument` API call.

As it is sadly the case for most `coq-lsp` features, I don't have the
resources to add a test case, so tagging this with `needs: test-case`
for the future.

Fixes #322