[fleche] [error] Detect Qed based on buffer contents
We do this for when the AST-based method doesn't work, either because
of parsing error, or because the `Qed` was parsed as a part of a
previous AST, like in `apply Qed.`.

Note that this happens because `Qed.` is not a keyword :p

This implies a huge improvement in usability when editing proofs.

Fixes #33