[lsp] Implement `textDocument/selectionRange`.
We return the range of the Coq sentence underlying the cursor. In
VSCode, this is triggered by the "Expand Selection" command.

Note the current implementation is partial: we only take into account
the first position, and we only return a single range (Coq sentence)
without parents.

This could be of help for #580.