[fleche] Store document environment in [Doc.t]
This is convenient as we may need to re-create the initial document
state when the environment has changed.

This patch doesn't fully take advantage of the document env, but it is
a prerequisite for the cleanup in #604, in particular to allow a
document to recover on `bump` if the init state failed.