[fleche] Experimental option for "lazy" document checking.
A new setting `check_only_on_request`, will instruct Flèche to only
check documents where data has been demanded.

This doesn't work well with some LSP calls, like `documentSymbols`, as
they require the whole document to be processed; if we would try to
serve those, we would always proceed to check the documents anyway,
defeating the purpose of this PR.

We thus refuse to serve these requests when in this mode.

This setting is just for experimentation, as it is not ready for
general, but likely will help people willing to add different
navigation strategies as discussed in Zulip.

cc #24