[fleche] Handle performance data correctly for cached sentences
We extend `Memo` as to store the original performance execution data,
this way, we can display it correctly, and we distinguish the case
where a sentence was memoized better.

This is a fix, because before we would display the wrong data for
incremental checking.

This is complementary to #686 and #689.

We cherry pick the protocol fixes from #689 as to have this PR in
mergeable / testeable shape.

We also send by default the full document perf data, instead of the
atop 10 hotspots.