[code] Display and allow status bar item to toggle continuous / lazy mode
We display the continuous / lazy mode checking in the status bar item,
and we allow to toggle them with a click.

There are some usability questions with the current setup tho, I
wonder if we should move the start / stop server command to the
tooltip or to a `LanguageStatusItem` as done in other extensions.

cc: #24