[code] Don't show goals type when details are closed.
This is a workaround for #525 #652

For now, I disable this setting, so types are not displayed.

The right fix would be to add a checkbox in the view, so this can be
toggled dynamically, but I don't have the cycles for that right now,
and not worth delaying the release which is long overdue.