[code] Show server status using LanguageStatusItem
We show server checking status and toolchain version using the new
`LanguageStatusItem` UI. It is also possible to control using this UI
whether the server is in continuous or lazy mode. Some more changes:

- Added a new command to toggle the mode.
- Show `coq-lsp` version info.

More information about what the server is doing should be possible if
the users want! (For example, we should show memory use, etc...)