[goals] Allow `command` to process several Coq commands.
This is a prototype extension.

Note the important change here in behavior here, but IMHO it is
needed: if the commands in `command` fail, we will now fail the
request.

I think in general we may allow some form of error recovery here, but
for now we just let clients recover.

It does seem that CoqPilot / Pytanque do need even a more powerful
interface, but this will allow the them to make some progress for nwo.

Thanks to CoqPilot devs for feedback on this.