Stop relying on environ internals.
More than half of the fields exposed in the private Environ.env module
are not supposed to be accessed by anybody outside of the implementation.
Some are actually just serialized to opaque data by Serlib.

Instead of exposing all this crap to the serialization process, we go
through a better behaved type that only exposes the bits that are supposed
to be accessed.