[controller] [lsp] Include Rocq feedback on request errors.
We use the optional `data` field from the LSP specification.

This is useful to be able to obtain feedback messages such as debug
messages even when a request fails.

This also opens the door to better protocol handling and petanque
integration.

We don't't update the VSCode to use this yet, but maybe we should do
it soon.