[lsp] [fleche] New notification `$/coq/executionInformation`
Flèche will emit `$/coq/executionInformation` after parsing a
sentence (tho we now its range) but _before_ to start executing it.

This is useful for clients, change was suggested by Josselin Poiret.

A key question remains here tho, and it is about the relationship of
this notification with `coq/fileProgress`.

It seems to me that a better way would be to update `coq/fileProgress`
as to reflect different status of the file processing architecture, in
particular, we could have a finer grained processing status.